uplink: read encryption key from separated file
The encryption key for encrypting the data has been moved out from the
configuration file to a separated file hence the uplink configuration
doesn't contain anymore the key, but it contains the path of the file
which contains the key.

This commit removed the encryption key from the configuration struct and
added a new field which contains the path to the encryption key
filepath with a method that allows to retrieve the key from it.